+++ src/vues/backoffice/vues/users/partner-users/dialog/user-details/partner-user-details-dialog.html
@@ -2,7 +2,7 @@
     v-bind:show="visible"
     v-bind:modal-class="'modal-medium'"
     v-on:close-dialog="close">
-    <h4 slot="header">Lottery User Details</h4>
+    <h4 slot="header">Partner User Details</h4>
     <div slot="body">
         <basic-user-details v-bind:user="user" />
     </div>
